backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
Flask-Swagger-UI==4.11.1
flasgger==0.9.7.1
cloudinary==1.34.0
python-dotenv==1.0.1
psycopg2-binary
requests==2.31.0
gunicorn==21.2.0
Werkzeug==2.3.8

# Dev/Test
pytest==7.4.0
pytest-flask==1.3.0
